renamed sim to input
