+++ mmelemental/models/input/dynamics.py
@@ -4,34 +4,27 @@ from pydantic import Field
 from typing import Any, Tuple, Union, Optional
 from qcelemental.models.types import Array
 
-class Dynamics(SimBase):
+class DynamicsInput(SimBase):
     """ Molecular dynamics parameter input schema."""
+
+    # System fields
+    mol: Union[Tuple[Molecule], Molecule] = Field(..., description = 'Molecular mechanics molecule object(s).')
 
     # Velocity fields
     gen_vel: Array[bool] = Field(None, description = 'Generate velocities from Maxwell distribution(s) at temperature(s) gen_temp, with random seed(s) gen_seed.')
     gen_temp: Array[float] = Field(None, description = 'Temperature(s) in Kelvin for Maxwell distribution(s) used to generate gen_vel.')
     gen_seed: Array[int] = Field(None, description = 'Used to initialize random generator for random velocities for each group.')
 
-    # Thermostat fields
+    # Temperature coupling fields
     temp: Array[float] = Field(None, description = 'Temperature(s) for thermostat(s) in Kelvin.')
+    temp_method: str
     
-    # Barostat fields
-    press_couple: str = Field(None, description = 'Pressure coupling method.')
+    # Pressure coupling fields
+    press_method: str = Field(None, description = 'Pressure coupling method.')
+    press: Array[float] = Field(None, description = 'Pressure(s) used for barostate in bar. The number of required values is implied by press_couple_type.')
     press_couple_type: str = Field(None, description = 'Specifies the type of isotropy used for pressure coupling.')
     press_couple_freq: int = Field(None, description = 'Frequency used for coupling the pressure.')
-    press_couple_relax: float = Field(None, description = 'The relaxation time constant in picoseconds used for pressure coupling (one value for all directions?).')
     compressibility: Array[float] = Field(None, description = 'Compressibility factor(s) in 1/bar. The number of required values is implied by press_couple_type.')
-    press: Array[float] = Field(None, description = 'Pressure(s) used for barostate in bar. The number of required values is implied by press_couple_type.')
-    
-    # NAMD fields
-    use_group_press: Optional[bool] = Field(None, description = 'Pressure is calculated using either the atomic virial and kinetic energy (the default) or a hydrogen-group based pseudo-molecular virial and kinetic energy')
-    use_flex_cell: Optional[bool] = Field(None, description = 'Allows the three orthogonal dimensions of the periodic cell to fluctuate independently when this option is enabled.')
-    use_const_ratio: Optional[bool] = Field(None, description = 'When enabled, the ratio of unit cell in the xy plane is kept constant while allowing fluctuations along all axes.')
-    use_const_area: Optional[bool] = Field(None, description = 'When enabled, the ddimension of the unit cell in the xy is kept plane constant while allowing fluctuations along the z axis.')
-    #LangevinPistonTemp 
-    #SurfaceTensionTarget 
-    #StrainRate 
-    #ExcludeFromPressure 
     
     # GROMACS fields
     press_scale: str = Field(None, description = 'Which reference coordinates to scale the matrix of the pressure coupling with. Options: no, all, or com.')
