+++ mmelemental/models/input/sim_writer.py
@@ -1,5 +1,5 @@
 from mmelemental.models.base import Base
-from mmelemental.models.sim.base import SimBase
+from mmelemental.models.input.base import SimBase
 from typing import List, Optional, Union, Dict
 from pydantic import Field
 from typing import Dict, List, Any, Optional, Tuple
