--- mmelemental/models/sim/__init__.py
@@ -1,1 +1,0 @@
-from . import *
